requirements.txt
torch>=2.1.0
numpy>=1.26.4
Pillow>=10.0.0